Test and update PIO program
Simulated PIO program with https://github.com/soundpaint/rp2040pio
Unfortunately MOV doesn't trigger autopush, so the CPU will have to invert the result itself.
The OUT NULL 32 and MOV Y ! NULL commands can be executed in the C program to save a couple of instructions. Not that anything else is going to use the PIO instruction memory anyway...